requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.3=py311h2bbff1b_0
aiosignal=1.3.1=pyhd8ed1ab_0
anyio=3.7.1=pypi_0
argilla=1.13.3=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
backcall=0.2.0=pyh9f0ad1d_0
backoff=2.2.1=pypi_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.2=py311haa95532_0
blas=1.0=mkl
brotli-python=1.0.9=py311hd77b12b_7
bs4=4.12.2=py38hd3eb1b0_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
certifi=2023.7.22=py311haa95532_0
cffi=1.15.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
chroma-hnswlib=0.7.2=pypi_0
chromadb=0.4.5=pypi_0
click=8.1.6=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
comm=0.1.4=pyhd8ed1ab_0
cryptography=41.0.3=pypi_0
dataclasses-json=0.5.7=pyhd8ed1ab_0
debugpy=1.6.7=py311hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pypi_0
docx2txt=0.8=pypi_0
et-xmlfile=1.1.0=pypi_0
executing=1.2.0=pyhd8ed1ab_0
fastapi=0.99.1=pypi_0
filelock=3.12.2=pypi_0
filetype=1.2.0=pypi_0
flatbuffers=23.5.26=pypi_0
frozenlist=1.3.3=py311h2bbff1b_0
fsspec=2023.6.0=pypi_0
greenlet=2.0.1=py311hd77b12b_0
h11=0.14.0=pypi_0
httpcore=0.16.3=pypi_0
httptools=0.6.0=pypi_0
httpx=0.23.3=pypi_0
huggingface-hub=0.16.4=pypi_0
humanfriendly=10.0=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib-resources=6.0.1=pypi_0
importlib_metadata=6.8.0=hd8ed1ab_0
instructorembedding=1.0.1=pypi_0
intel-openmp=2023.2.0=h57928b3_49496
ipykernel=6.25.1=pyh6817e22_0
ipython=8.14.0=pyh08f2357_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h1ea47a8_0
langchain=0.0.239=pyhd8ed1ab_0
langsmith=0.0.20=pyhd8ed1ab_0
libffi=3.4.4=hd77b12b_0
libsodium=1.0.18=h8d14728_1
lxml=4.9.3=pypi_0
markdown=3.4.4=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.20.1=pyhd8ed1ab_0
marshmallow-enum=1.5.1=pyh9f0ad1d_3
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.2.0=h6a75c08_49573
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.6=py311hf62ec03_1
mkl_random=1.2.2=py311hf62ec03_1
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
msg-parser=1.2.0=pypi_0
multidict=6.0.2=py311h2bbff1b_0
mypy_extensions=1.0.0=pyha770c72_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
nltk=3.8.1=pypi_0
numexpr=2.8.4=py311h1fcbade_1
numpy=1.23.5=pypi_0
olefile=0.46=pypi_0
onnxruntime=1.15.1=pypi_0
openai=0.27.8=pypi_0
openapi-schema-pydantic=1.2.4=pyhd8ed1ab_0
openpyxl=3.1.2=pypi_0
openssl=3.0.10=h2bbff1b_0
overrides=7.4.0=pypi_0
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pdf2image=1.16.3=pypi_0
pdfminer-six=20221105=pypi_0
pickleshare=0.7.5=py_1003
pillow=10.0.0=pypi_0
pip=23.2.1=py311haa95532_0
platformdirs=3.10.0=pyhd8ed1ab_0
posthog=3.0.1=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.24.0=pypi_0
psutil=5.9.0=py311h2bbff1b_0
pulsar-client=3.2.0=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pypi_0
pydantic=1.10.8=py311h2bbff1b_0
pygments=2.16.1=pyhd8ed1ab_0
pypandoc=1.11=pypi_0
pypika=0.48.9=pypi_0
pyreadline3=3.4.1=pypi_0
pysocks=1.7.1=pyh0701188_6
python=3.11.4=he1021f5_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-docx=0.8.11=pypi_0
python-dotenv=1.0.0=pypi_0
python-magic=0.4.27=pypi_0
python-pptx=0.6.21=pypi_0
python_abi=3.11=2_cp311
pytz=2023.3=pypi_0
pywin32=305=py311h2bbff1b_0
pyyaml=6.0=py311h2bbff1b_1
pyzmq=25.1.0=py311hd77b12b_0
regex=2023.8.8=pypi_0
requests=2.31.0=pyhd8ed1ab_0
rfc3986=1.5.0=pypi_0
rich=13.5.2=pypi_0
safetensors=0.3.2=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.1=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py311haa95532_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.4=py311haa95532_0
sqlalchemy=1.4.39=py311h2bbff1b_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.27.0=pypi_0
stringcase=1.2.0=py_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.2=pyhd8ed1ab_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.4.0=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.13.3=pypi_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.3.2=py311h2bbff1b_0
tqdm=4.66.1=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.31.0=pypi_0
typer=0.7.0=pypi_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
typing_inspect=0.9.0=pyhd8ed1ab_0
tzdata=2023c=h04d1e81_0
unstructured=0.7.12=pypi_0
urllib3=2.0.4=pyhd8ed1ab_0
uvicorn=0.23.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchfiles=0.19.0=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
websockets=11.0.3=pypi_0
wheel=0.38.4=py311haa95532_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wrapt=1.14.1=pypi_0
xlrd=2.0.1=pypi_0
xlsxwriter=3.1.2=pypi_0
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=h8ffe710_2
yarl=1.8.1=py311h2bbff1b_0
zeromq=4.3.4=h0e60522_1
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
